TS Linting and DI support for the Merkle project
MerkleModule now supports a forFeature() method that accepts a
ModuleIdentifier, MerkleTreeDescription, and optinal Tag discriminator.

It creates a DynamicModule that has a unique set of tokens for the suite
of MerkleTree utility classes derived from the given
MerkleTreeDescription (a Calculator, LocatorFactory, and a
CanonicalPathNamer).

Also pruned an unused interface from the package--some spring cleaning.